backend/pyproject.toml
[project]
name = "f1-journey-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "rapidfuzz>=3.0.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]
